shorthand flag syntax
