+++ packages/bot-utils/src/index.ts
@@ -1,6 +1,7 @@
 export * from './flags';
 export * from './get';
 export * from './logger';
+export * from './measurements-request';
 export * from './post';
 export * from './types';
 export * from './utils';
