+++ packages/bot-utils/tests/measurements-request.test.ts
@@ -1,13 +1,14 @@
 import { describe, expect, it } from 'vitest';
 
-import { argsToFlags, parseFlags } from '../src/flags';
+import { argsToFlags } from '../src/flags';
+import { buildPostMeasurements } from '../src/measurements-request';
 
 describe('Utils', () => {
 	describe('parse arguments', () => {
 		describe('ping', () => {
 			it('should parse ping args', () => {
 				const args = 'ping google.com from New York';
-				const result = parseFlags(argsToFlags(args));
+				const result = buildPostMeasurements(argsToFlags(args));
 				expect(result).toEqual([{
 					type: 'ping',
 					target: 'google.com',
@@ -20,7 +21,7 @@ describe('Utils', () => {
 
 			it('should parse ping args with flags', () => {
 				const args = 'ping google.com from New York --limit 2 --packets 3';
-				const result = parseFlags(argsToFlags(args));
+				const result = buildPostMeasurements(argsToFlags(args));
 				expect(result).toEqual([{
 					type: 'ping',
 					target: 'google.com',
@@ -35,14 +36,14 @@ describe('Utils', () => {
 
 			it('should throw if incorrect ping flag', () => {
 				const args = 'ping google.com from New York --limit 2 --packets 3 --protocol icmp';
-				expect(() => parseFlags(argsToFlags(args))).toThrow('Invalid option "protocol" for "ping"!\nExpected "packets, latency, target, from, limit".');
+				expect(() => buildPostMeasurements(argsToFlags(args))).toThrow('Invalid option "protocol" for "ping"!\nExpected "packets, latency, target, from, limit".');
 			});
 		});
 
 		describe('traceroute', () => {
 			it('should parse traceroute args', () => {
 				const args = 'traceroute google.com from New York';
-				const result = parseFlags(argsToFlags(args));
+				const result = buildPostMeasurements(argsToFlags(args));
 				expect(result).toEqual([{
 					type: 'traceroute',
 					target: 'google.com',
@@ -55,7 +56,7 @@ describe('Utils', () => {
 
 			it('should parse traceroute args with flags', () => {
 				const args = 'traceroute google.com from New York --limit 2 --protocol tcp --port 80';
-				const result = parseFlags(argsToFlags(args));
+				const result = buildPostMeasurements(argsToFlags(args));
 				expect(result).toEqual([{
 					type: 'traceroute',
 					target: 'google.com',
@@ -71,14 +72,14 @@ describe('Utils', () => {
 
 			it('should throw if incorrect traceroute flag', () => {
 				const args = 'traceroute google.com from New York --limit 2 --protocol tcp --port 80 --packets 3';
-				expect(() => parseFlags(argsToFlags(args))).toThrow('Invalid option "packets" for "traceroute"!\nExpected "protocol, port, target, from, limit".');
+				expect(() => buildPostMeasurements(argsToFlags(args))).toThrow('Invalid option "packets" for "traceroute"!\nExpected "protocol, port, target, from, limit".');
 			});
 		});
 
 		describe('dns', () => {
 			it('should parse dns args', () => {
 				const args = 'dns google.com from New York';
-				const result = parseFlags(argsToFlags(args));
+				const result = buildPostMeasurements(argsToFlags(args));
 				expect(result).toEqual([{
 					type: 'dns',
 					target: 'google.com',
@@ -91,7 +92,7 @@ describe('Utils', () => {
 
 			it('should parse dns args with flags', () => {
 				const args = 'dns google.com from New York --limit 2 --query AAAA --protocol tcp --port 80 --resolver 1.1.1.1 --trace';
-				const result = parseFlags(argsToFlags(args));
+				const result = buildPostMeasurements(argsToFlags(args));
 				expect(result).toEqual([{
 					type: 'dns',
 					target: 'google.com',
@@ -112,14 +113,14 @@ describe('Utils', () => {
 
 			it('should throw if incorrect dns flag', () => {
 				const args = 'dns google.com from New York --limit 2 --query AAAA --protocol tcp --port 80 --test';
-				expect(() => parseFlags(argsToFlags(args))).toThrow('Invalid option "test" for "dns"!\nExpected "query, protocol, port, resolver, trace, latency, target, from, limit".');
+				expect(() => buildPostMeasurements(argsToFlags(args))).toThrow('Invalid option "test" for "dns"!\nExpected "query, protocol, port, resolver, trace, latency, target, from, limit".');
 			});
 		});
 
 		describe('mtr', () => {
 			it('should parse mtr args', () => {
 				const args = 'mtr google.com from New York';
-				const result = parseFlags(argsToFlags(args));
+				const result = buildPostMeasurements(argsToFlags(args));
 				expect(result).toEqual([{
 					type: 'mtr',
 					target: 'google.com',
@@ -132,7 +133,7 @@ describe('Utils', () => {
 
 			it('should parse mtr args with flags', () => {
 				const args = 'mtr google.com from New York --limit 2 --protocol tcp --port 80 --packets 16';
-				const result = parseFlags(argsToFlags(args));
+				const result = buildPostMeasurements(argsToFlags(args));
 				expect(result).toEqual([{
 					type: 'mtr',
 					target: 'google.com',
@@ -149,14 +150,14 @@ describe('Utils', () => {
 
 			it('should throw if incorrect mtr flag', () => {
 				const args = 'mtr google.com from New York --limit 2 --protocol tcp --port 80 --packets 16 --test';
-				expect(() => parseFlags(argsToFlags(args))).toThrow('Invalid option "test" for "mtr"!\nExpected "protocol, port, packets, target, from, limit".');
+				expect(() => buildPostMeasurements(argsToFlags(args))).toThrow('Invalid option "test" for "mtr"!\nExpected "protocol, port, packets, target, from, limit".');
 			});
 		});
 
 		describe('http', () => {
 			it('should parse http args', () => {
 				const args = 'http google.com from New York';
-				const result = parseFlags(argsToFlags(args));
+				const result = buildPostMeasurements(argsToFlags(args));
 				expect(result).toEqual([{
 					type: 'http',
 					target: 'google.com',
@@ -171,7 +172,7 @@ describe('Utils', () => {
 
 			it('should parse http args with flags', () => {
 				const args = 'http google.com from New York --limit 2 --path / --query ?a=abc --host google.com --method get --port 80 --protocol https --header Content-Type: text/html; charset=utf-8 --header Content-Encoding: gzip';
-				const result = parseFlags(argsToFlags(args));
+				const result = buildPostMeasurements(argsToFlags(args));
 
 				expect(result).toEqual([{
 					type: 'http',
@@ -198,7 +199,7 @@ describe('Utils', () => {
 
 			it('should infer url for http parse', () => {
 				const args = 'http https://google.com:80/test?a=abc from New York --limit 2 --method get --header Content-Encoding: gzip --header Content-Type: text/html; charset=utf-8';
-				const result = parseFlags(argsToFlags(args));
+				const result = buildPostMeasurements(argsToFlags(args));
 
 				expect(result).toEqual([{
 					type: 'http',
@@ -225,7 +226,7 @@ describe('Utils', () => {
 
 			it('should throw if incorrect http flag', () => {
 				const args = 'http google.com from New York --limit 2 --path / --query ?a=abc --host google.com --method get --port 80 --protocol https --header Content-Encoding: gzip --header Content-Type: text/html; charset=utf-8 --test';
-				expect(() => parseFlags(argsToFlags(args))).toThrow('Invalid option "test" for "http"!\nExpected "protocol, port, method, path, query, host, header, latency, target, from, limit".');
+				expect(() => buildPostMeasurements(argsToFlags(args))).toThrow('Invalid option "test" for "http"!\nExpected "protocol, port, method, path, query, host, header, latency, target, from, limit".');
 			});
 		});
 	});
